refactor: bilibili-share-to-friends 规范组件文件命名
