+++ scripts/bilibili-share-to-friends/src/components/dialog-footer/DialogFooter.jsx
