+++ scripts/bilibili-share-to-friends/src/components/dialog-header/DialogHeader.jsx
