+++ scripts/bilibili-share-to-friends/src/components/dialog/Dialog.jsx
