+++ scripts/bilibili-share-to-friends/src/components/entry-button/EntryButton.jsx
