+++ scripts/bilibili-share-to-friends/src/components/index.js
@@ -1,0 +1,11 @@
+export * from "./dialog/Dialog.jsx";
+export * from "./dialog-footer/DialogFooter.jsx";
+export * from "./dialog-header/DialogHeader.jsx";
+export * from "./entry-button/EntryButton.jsx";
+export * from "./recipient-tabs/RecipientTabs.jsx";
+export * from "./relation-filter/RelationFilter.jsx";
+export * from "./search-box/SearchBox.jsx";
+export * from "./state-view/StateView.jsx";
+export * from "./user-list/UserList.jsx";
+export * from "./user-list-item/UserListItem.jsx";
+export * from "./video-preview/VideoPreview.jsx";
