+++ scripts/bilibili-share-to-friends/src/components/recipient-tabs/style.css
