+++ scripts/bilibili-share-to-friends/src/components/search-box/SearchBox.jsx
