+++ scripts/bilibili-share-to-friends/src/components/state-view/StateView.jsx
