+++ scripts/bilibili-share-to-friends/src/components/user-list-item/UserListItem.jsx
