+++ scripts/bilibili-share-to-friends/src/components/video-preview/VideoPreview.jsx
