+++ scripts/bilibili-share-to-friends/src/components/video-preview/style.css
